RAG/requirements.txt
# python version 3.10
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
loguru==0.7.3
oci==2.141.0
langchain-community==0.3.12
python-multipart==0.0.20

# Database
oracledb==2.5.1

# File handling
python-docx==1.1.2
pypdf2==3.0.1
pypdf==5.1.0
chardet==5.2.0

# Temporary
chromadb==0.5.23
cohere==5.13.3

# Testing
pytest==8.1.1
requests==2.31.0
pandas==2.2.1
openpyxl==3.1.2